fibonacci to 50

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,9 +7,9 @@
 int main(void)
 {
 	int i;
-	int j;
-	int k;
-	int res;
+	long int j;
+	long int k;
+	long int res;
 
 	j = 1;
 	k = 2;
